Move test function into header and template it
